requirements.txt
argon2-cffi==21.1.0
asgiref==3.5.2
astroid==2.8.0
async-generator==1.10
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
backcall==0.2.0
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
bleach==4.1.0
blis==0.7.9
catalogue==2.0.8
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
confection==0.0.3
coverage==5.5
coveralls==3.2.0
cryptography==36.0.2
cycler==0.11.0
cymem==2.0.7
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
distlib==0.3.3
Django==4.1.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
docopt==0.6.2
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.4.0/en_core_web_lg-3.4.0-py3-none-any.whl
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.1.0
filetype==1.0.13
Flask==2.2.2
fonttools==4.28.3
h11==0.13.0
idna==3.2
importlib-metadata==5.0.0
iniconfig==1.1.1
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
isort==5.9.3
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
jsonschema==4.1.2
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.2
langcodes==3.3.0
lazy-object-proxy==1.6.0
MarkupSafe==2.1.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
mccabe==0.6.1
mistune==0.8.4
murmurhash==1.0.9
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.5
notebook==6.4.5
numpy==1.21.2
opencv-python==4.5.5.64
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
pathy==0.6.2
pdf2image==1.16.0
pickleshare==0.7.5
Pillow==9.1.1
pipreqs==0.4.10
platformdirs==2.3.0
pluggy==1.0.0
preshed==3.0.8
prometheus-client==0.12.0
prompt-toolkit==3.0.21
psycopg2-binary==2.9.5
py==1.10.0
pycodestyle==2.7.0
pycosat==0.6.3
pycparser==2.20
pydantic==1.9.2
Pygments==2.10.0
pylint==2.11.1
PyMuPDF==1.19.6
pyOpenSSL==22.0.0
pyparsing==2.4.7
PyPDF2==1.28.2
pyphen==0.13.0
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
PyQt5-stubs==5.15.2.0
pyrsistent==0.18.0
PySocks==1.7.1
pytesseract==0.3.9
pytest==6.2.5
pytest-cov==3.0.0
python-dateutil==2.8.2
pytz==2021.1
#pywin32==302
#pywinpty==1.1.5
pyzmq==22.3.0
regex==2021.11.10
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.1
seaborn==0.12.1
selenium==4.1.3
Send2Trash==1.8.0
setuptools-scm==6.3.2
six==1.16.0
sklearn==0.0
smart-open==5.2.1
sniffio==1.2.0
sortedcontainers==2.4.0
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
sqlparse==0.4.3
srsly==2.4.5
terminado==0.12.1
testpath==0.5.0
textblob==0.17.1
textstat==0.7.3
thinc==8.1.5
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
trio==0.20.0
trio-websocket==0.9.2
typer==0.4.2
typing-extensions==3.10.0.2
tzdata==2022.6
urllib3==1.26.7
virtualenv==20.8.1
wasabi==0.10.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.12.1
wsproto==1.1.0
yarg==0.1.9
zipp==3.10.0
